Refactoring of RoxyFileman functions
(cherry picked from commit 1967af1432d100097f255b8d49a2ad5d65b1bfa4)
